# inkwell_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 gin + MySQL + Redis + JWT 的社区后端接口文档。\n所有接口返回统一结构: {\"code\":1000,\"message\":\"success\",\"data\":{}}。业务失败时 HTTP 状态码仍然是 200(只有访问不存在的路由才返回 404), 需要根据 code 判断结果, code 取值见 controller/code.go。\n需要登录的接口要在请求头带上 Authorization: Bearer \u003caccessToken\u003e; access token 失效(1006)时可以调用 /api/v1/refresh_token 刷新后重放请求。\n所有 64 位 ID(用户/帖子/评论)在 JSON 中都使用字符串传输, 用数字传会丢精度。",
        "title": "Inkwell 社区 API",
        "contact": {},
        "version": "0.1.1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/login": {
            "post": {
                "description": "登录成功后返回 accessToken(2 小时有效)与 refreshToken(7 天有效)。\n用户名不存在和密码错误都返回 code 1004, 避免被用来枚举用户名。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/refresh_token": {
            "get": {
                "description": "access token 放在 Authorization 请求头(已过期也要带), refresh token 放在 query 参数里。\ntoken 无效或已过期时返回 code 1006。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "刷新 access token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccessToken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "refresh token",
                        "name": "refresh_token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.RefreshTokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "注册新用户。用户名长度 2-64, 密码长度 6-64, 两次密码必须一致。\n失败时 HTTP 状态码仍是 200, 通过响应体里的 code 判断: 参数错误 1001, 用户名重复 1002。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册参数, confirm_password 和 re_password 传一个即可",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功, data 为 null",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseData"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "查询用户的公开信息(用户名、注册时间), 游客可访问。用户不存在或 ID 非法时返回 code 1009。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户主页",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID(雪花 ID, 用字符串传)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.UserProfile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/{id}/post": {
            "get": {
                "description": "分页查询某个用户发布的帖子(按发帖时间倒序), 游客可访问, 票数与评论数来自 Redis。\n用户不存在时返回 code 1009, ID 非法时返回 code 1001。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户发布的帖子",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID(雪花 ID, 用字符串传)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码, 从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页条数, 最大50",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "帖子列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.PostListItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.MyCode": {
            "type": "integer",
            "format": "int64",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006,
                1007,
                1008,
                1009,
                1010
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeInvalidParams",
                "CodeUserExist",
                "CodeUserNotExist",
                "CodeInvalidPassword",
                "CodeServerBusy",
                "CodeInvalidToken",
                "CodeInvalidAuthFormat",
                "CodeNotLogin",
                "CodeNotFound",
                "CodeTooManyRequests"
            ]
        },
        "controller.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.MyCode"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.LoginForm": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "zhangsan"
                }
            }
        },
        "models.LoginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "userID": {
                    "description": "UserID 是雪花算法生成的 64 位 ID, 用字符串返回避免前端精度丢失",
                    "type": "string",
                    "example": "3458764513820540928"
                },
                "username": {
                    "type": "string",
                    "example": "zhangsan"
                }
            }
        },
...
        "models.PostListItem": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "0"
                },
                "author_name": {
                    "type": "string"
                },
                "comment_num": {
                    "type": "integer"
                },
                "community_id": {
                    "type": "integer"
                },
                "community_name": {
                    "type": "string"
                },
                "create_time": {
                    "description": "发帖时间(unix 秒)",
                    "type": "integer"
                },
                "post_id": {
                    "type": "string",
                    "example": "0"
                },
                "score": {
                    "description": "榜单排序分, 只有榜单接口会返回",
                    "type": "number"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "vote_num": {
                    "description": "VoteNum 净票数(赞成 - 反对), 可能为负",
                    "type": "integer"
                }
            }
        },
        "models.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "models.RegisterForm": {
            "type": "object",
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "re_password": {
                    "description": "RePassword 兼容旧版前端使用的 re_password 字段",
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserProfile": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "0"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "取值格式 \"Bearer \u003caccessToken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}